% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBSLoader.R
\name{pullBBS}
\alias{pullBBS}
\title{Pull data from the North American Breeding Bird Survey ftp server}
\usage{
pullBBS(year, country, region, AOU, useCache = F)
}
\arguments{
\item{year}{A vector containing a list of the years for the surveys}

\item{country}{ISO country code for the region to pull observations from.  Use the function pullBBSmeta to get a list of country and region ISO codes.}

\item{region}{A BBS region code representing a country or state.  Use the function pullBBSmeta to get a list of country and region codes.  Can be in the form of a list for selecting multiple regions.  If no region is specified, all regions within the country will be pulled.}

\item{AOU}{Vector of AOU species codes to pull.  AOU = 0 will pull all species within the sepified region.  If no AOU codes are specified, all species observed within the regions for the given period will be pulled.}

\item{useCache}{When set to TRUE, downloaded data will not be erased after the function is complete.  In subsequent runs, data will be loaded taken from the cache instead of being downloaded.  This is usefull if you are running the function multiple times in a session and you don't want to redownload the BBS data on each run.}
}
\value{
A data.frame containing all the selected routes and observations for the specified species.
}
\description{
This will automatically download all the BBS data from the FTP server.  Only those routes within your specified regions and sightings from your specified species will be saved in the output data.frame.  This function will also use the weather data from the BBS server to create missing absence records for routes in which your species was not observed.
}

