% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBSLoader.R
\name{pullBBSmeta}
\alias{pullBBSmeta}
\title{Pull data from the North American Breeding Bird Survey ftp server}
\usage{
pullBBSmeta()
}
\description{
This will create a folder called BBS_Meta in your working directory where the metadata files from the BBS server will be saved to.  You can use this metadata to find out mode information about the sampling protocol and species and country/region codes.
}

