% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBSLoader.R
\name{pullBBSRouteLocations}
\alias{pullBBSRouteLocations}
\title{Pull data from the North American Breeding Bird Survey ftp server}
\usage{
pullBBSRouteLocations(country, region, ignoreCache = F)
}
\arguments{
\item{country}{A BBS country code.  Use the function pullBBSmeta to get a list of country and region codes.}

\item{region}{A BBS region code representing a country or state.  Use the function pullBBSmeta to get a list of country and region codes.  Can be in the form of a list for selecting multiple regions.}

\item{ignoreCache}{Set to TRUE to overwrite previously downloaded data.  This is usefull if a prevously cached file has been corrupted due to a download error.}
}
\value{
A data.frame containing all the selected routes and their lat/long coordinates.
}
\description{
Pull data from the North American Breeding Bird Survey ftp server
}

