% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBSLoader.R
\name{WideToLongBBS}
\alias{WideToLongBBS}
\title{Convert BBS data into a long format}
\usage{
WideToLongBBS(routes)
}
\arguments{
\item{routes}{A data.frame returned form the pullBBS function.}
}
\description{
When used on the data.frame returned by pullBBS, the 50 columns holding observations for each of the 50 stops will be condensed into two columns: 'Stop' and 'Observations'.
}

