% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBSLoader.R
\name{pullBBS}
\alias{pullBBS}
\title{Pull data from the North American Breeding Bird Survey ftp server}
\usage{
pullBBS(year, country, region, AOU, ignoreCache = F)
}
\arguments{
\item{year}{A vector containing a list of the years for the surveys}

\item{country}{A BBS country code.  Use the function pullBBSmeta to get a list of country and region codes.}

\item{region}{A BBS region code representing a country or state.  Use the function pullBBSmeta to get a list of country and region codes.  Can be in the form of a list for selecting multiple regions.}

\item{AOU}{AOU species code.  Use pullBBS meta to get a species list with codes.}

\item{ignoreCache}{Set to TRUE to overwrite previously downloaded data.  This is usefull if a prevously cached file has been corrupted due to a download error..}
}
\value{
A data.frame containing all the selected routes and observations for the specified species.
}
\description{
This will automatically download all the BBS data from the FTP server.  Only those routes within your specified regions and sightings from your specified species will be saved in the output data.frame.  This function will also use the weather data from the BBS server to create missing absence records for routes in which your species was not observed.
}

